requirements-dev.txt
pygments
sphinx
sphinx-autobuild
coverage
django-generic-helpers
django-mptt
docutils
flake8
tox
wheel
